first draft for LuceneQuery that used lucene api
The DirectApiLuceneQuery (yeah, the name is bad, will probably change later)
directly accesses the lucene API (that means it constructs terms and combines them).
It appears to be quite compatible with my interface, which is good.
I just have to sort out how to use these analyzers outside of QueryParser ^^.